# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DailyCountView": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "count"
        ],
        "title": "DailyCountView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequestStatisticsView": {
        "properties": {
          "daily_requests": {
            "items": {
              "$ref": "#/components/schemas/DailyCountView"
            },
            "title": "Daily Requests",
            "type": "array"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "starter_requests": {
            "items": {
              "$ref": "#/components/schemas/StarterCountView"
            },
            "title": "Starter Requests",
            "type": "array"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "daily_requests",
          "starter_requests"
        ],
        "title": "RequestStatisticsView",
        "type": "object"
      },
...
      "StarterCountView": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "starter_id": {
            "title": "Starter Id",
            "type": "string"
          }
        },
        "required": [
          "starter_id",
          "count"
        ],
        "title": "StarterCountView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CultureChamp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/request-statistics": {
      "get": {
        "operationId": "request_statistics_api_v1_admin_request_statistics_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatisticsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Statistics",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}